% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sleep_params.R
\name{fetch_sleep_params}
\alias{fetch_sleep_params}
\title{Calculate sleep micro architecture parameters}
\usage{
fetch_sleep_params(hypnogram_file, epoch_duration)
}
\arguments{
\item{hypnogram_file}{A character array with sleep stages}

\item{epoch_duration}{Duration of sleep epoch in seconds}
}
\value{
A data frame with 32 columns
}
\description{
Calculate sleep micro architecture parameters
}
\examples{
sleep_data <- fetch_sleep_params(hypnogram_file = c('W' ,'W','N1','N2','N2','N3','R'),
epoch_duration = 30)
}
